Addition of header/banner, images and css
A banner at the top of the page with css and background images was
added. As well as attract backgrounds and logo’s added in an ‘img’
folder. HTML5 validity logo created and saved as SVG, optimized and
then base64 encoded so as to create a circular logo.
As well as ‘sitePaths.php’ file added to all folders and minor tweaking
of the header, footer and navBar.